requirements.txt
chromadb==0.5.5
fitz==0.0.1.dev2
GeneralAgent==0.3.25
ipython==8.12.3
matplotlib==3.9.2
ollama==0.3.3
openai==1.45.0
paddleocr==2.8.1
pymupdf4llm==0.0.14
python-dotenv==1.0.1
Requests==2.32.3
Shapely==2.0.6
pymupdf
torch  
paddlepaddle
numpy
fitz